devtools/client/shared/components/reps/test/mochitest/test_reps_long-string.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test LongString rep
-->
<head>
  <meta charset="utf-8">
  <title>Rep test - LongString</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript"></script>
<script type="application/javascript">
window.onload = Task.async(function* () {
  const {
    REPS,
    getRep,
  } = browserRequire("devtools/client/shared/components/reps/reps");
  let { LongStringRep } = REPS;

  try {
    // Test that correct rep is chosen
    is(getRep(getGripStub("testMultiline")), LongStringRep.rep,
      "Rep correctly selects LongString Rep");

    // Test rendering
    yield testMultiline();
    yield testMultilineOpen();
    yield testFullText();
    yield testMultilineLimit();
    yield testUseQuotes();
  } catch (e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }

  function quoteNewlines(text) {
    return text.split("\n").join("\\n");
  }

  function testMultiline() {
    const stub = getGripStub("testMultiline");
    const renderedComponent = renderComponent(
      LongStringRep.rep, { object: stub });

    is(renderedComponent.textContent, quoteNewlines(`"${stub.initial}…"`),
      "LongString rep has expected text content for multiline string");
  }

  function testMultilineLimit() {
    const renderedComponent = renderComponent(
      LongStringRep.rep, { object: getGripStub("testMultiline"), cropLimit: 20 });

    is(
      renderedComponent.textContent,
      `"a\\naaaaaaaaaaaaaaaaaa…"`,
      "LongString rep has expected text content for multiline string " +
      "with specified number of characters");
  }

  function testMultilineOpen() {
    const stub = getGripStub("testMultiline");
    const renderedComponent = renderComponent(
      LongStringRep.rep, { object: stub, member: {open: true}, cropLimit: 20 });

    is(renderedComponent.textContent, quoteNewlines(`"${stub.initial}…"`),
      "LongString rep has expected text content for multiline string when open");
  }

  function testFullText() {
    const stub = getGripStub("testFullText");
    const renderedComponentOpen = renderComponent(
      LongStringRep.rep, { object: stub, member: {open: true}, cropLimit: 20 });

    is(renderedComponentOpen.textContent, quoteNewlines(`"${stub.fullText}"`),
      "LongString rep has expected text content when grip has a fullText " +
      "property and is open");

    const renderedComponentNotOpen = renderComponent(
      LongStringRep.rep, { object: stub, cropLimit: 20 });

    is(renderedComponentNotOpen.textContent,
      `"a\\naaaaaaaaaaaaaaaaaa…"`,
      "LongString rep has expected text content when grip has a fullText " +
      "property and is not open");
  }

  function testUseQuotes() {
    const renderedComponent = renderComponent(LongStringRep.rep,
      { object: getGripStub("testMultiline"), cropLimit: 20, useQuotes: false });

    is(renderedComponent.textContent,
      "a\naaaaaaaaaaaaaaaaaa…",
      "LongString rep was expected to omit quotes");
  }

  function getGripStub(name) {
    const multilineFullText = "a\n" + Array(20000).fill("a").join("");
    const fullTextLength = multilineFullText.length;
    const initialText = multilineFullText.substring(0, 10000);

    switch (name) {
      case "testMultiline":
        return {
          "type": "longString",
          "initial": initialText,
          "length": fullTextLength,
          "actor": "server1.conn1.child1/longString58"
        };
      case "testFullText":
        return {
          "type": "longString",
          "fullText": multilineFullText,
          "initial": initialText,
          "length": fullTextLength,
          "actor": "server1.conn1.child1/longString58"
        };
    }
    return null;
  }
});
</script>
</pre>
</body>
</html>
